+++ packages/create-osmosis-app/types/cli.d.ts
@@ -1,1 +1,0 @@
-export declare const cli: (argv: any) => Promise<void>;