+++ examples/send-tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.2](https://github.com/cosmology-tech/create-cosmos-app/compare/@cosmology/send-tokens@1.8.1...@cosmology/send-tokens@1.8.2) (2023-07-14)
+
+**Note:** Version bump only for package @cosmology/send-tokens
+
+
+
+
+
 ## [1.8.1](https://github.com/cosmology-tech/create-cosmos-app/compare/@cosmology/send-tokens@1.8.0...@cosmology/send-tokens@1.8.1) (2023-06-28)
 
 **Note:** Version bump only for package @cosmology/send-tokens